# chore: cache members, mappable members, get attribute and optimize `CollectionInfoBu…