… src/Riok.Mapperly/Descriptors/WellKnownTypes.cs
… src/Riok.Mapperly/Helpers/SymbolExtensions.cs
… src/Riok.Mapperly/Symbols/MemberPath.cs