## Cached GetMembers, GetAccessibleMappableMembers and optimized `CollectionInfoBuilder`

### Description
Modified `WellKnownTypes` to cache more members, mappable members and attributes. Added a couple of checks to `CollectionInfoBuilder` to speed up common paths.

- Added cached methods to `WellKnownTypes` for `GetAttributes`, `GetAllMembers` and `GetAllAccessibleMappableMembers`
- CollectionTypeInfo
  - Converted `CollectionTypeInfo` to a struct
  - Removed duplicate `IImmutableQueue` and added `ImmutableSortedDictionary`
  - Added check to `HasValidAddMethod` for types with known `Add` methods, otherwise all none `None` types return false. `None` types are manually checked for `Add` methods.
  - `IsCountKnown` returns false for `IEnumerable` and true for all none `None` types. `None` types are manually checked for length properties.
  - Added a massive switch expression to `GetImplementedCollectionTypes`, it uses the known `CollectionTypeInfo` to get the implemented types. `None` types are manually checked for implemented types. Could be moved into `_collectionTypeInfos`.
    - Added a special check for `string` to return `IEnumerable`, because string isn't actually a "collection" type but does implement `IEnumerable` and is marked `None` by `GetCollectionTypeInfo`. It would iterate thrrought every possible collection type causing a considerable slow down.
- Moved `GetEnumeratedType` into `BuildCollectionInfos`. This is a quick check to ensure that both the source and target types are actually enumerable, saving a lot of wasted time.

I kept the slightly modified `GetAllMembers`and `GetAccessibleMappableMembers` (they call `WellKnownTypes`) in `SymbolExtensions`. Should I remove them and update all the call sites to use the `WellKnownTypes` version?

### Benchmarks

#### Optimized
|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.699 ms | 0.0322 ms | 0.0317 ms |   85.9375 |  19.5313 |  180.45 KB |
| LargeCompile | 54.299 ms | 0.9632 ms | 0.9460 ms | 1500.0000 | 375.0000 | 9344.24 KB |

#### Original
|       Method |       Mean |     Error |     StdDev |      Gen0 |     Gen1 |   Allocated |
|------------- |-----------:|----------:|-----------:|----------:|---------:|------------:|
|      Compile |   3.569 ms | 0.2264 ms |  0.6641 ms |  171.8750 |  31.2500 |   311.38 KB |
| LargeCompile | 100.170 ms | 4.6670 ms | 13.3905 ms | 2200.0000 | 800.0000 | 13395.01 KB |

Helps #72

## Checklist

- [x] The existing code style is followed
- [x] The commit message follows our guidelines
- [x] Performed a self-review of my code
- [x] Hard-to-understand areas of my code are commented
- [x] The documentation is updated (as applicable)
- [x] Unit tests are added/updated
- [x] Integration tests are added/updated (as applicable, especially if feature/bug depends on roslyn or framework version in use)