@latonz 2023-07-04
@TimothyMakkison thank you for this PR. Could you split it up into several smaller PR's to make them easier to review? This also makes sure you get prompter feedback and keeps the enhancements isolated. Additionally the performance impact of each change can be measured more easily. Proposed PR's:
* Caching of members
* CollectionTypeInfo general optimisations (struct, Immutable duplicate / sorted dict, GetEnumeratedType in BuildCollectionInfos)
* CollectionTypeInfo GetImplementedCollectionTypes

I'd split GetImplementedCollectionTypes and the general CollectionTypeInfo optimisations as I'm not sure if the massive type switch is worth it, would be nice if we could measure the performance impact on that separately and therefore create an isolated PR.

On a first glance the enhancements look promising 😊  I think I'd remove the member methods from `SymbolExtensions` and move them to the well known types class. Not sure if well known types is still an appropriate name then or if it should be renamed to just `Types` or similar.