@TimothyMakkison 2023-07-04
Good idea, I'll split this pr.

> I'm not sure if the massive type switch is worth it, 

It was probably the most impactful change. Iirc Mapperly spent 30% of the time calling GetImplementedCollections. It would also get the type symbol for every possible collection, degrading most use cases where they only simple collections.

I can remove the switch and move it into CollectionInfos as another property