@latonz 2023-07-04
@TimothyMakkison oh okay, I didn't expect that. I think I'd keep the current switch as it is easier to read 🤔
Doesn't this deactivate the feature detection? Eg. `IReadOnlySet` is only available for .NET 5+ and not for .NET Standard 2.0, this was detected before, but isn't in the new code. Not sure if this is an actual problem for Mapperly...