@TimothyMakkison 2023-07-04
> @TimothyMakkison oh okay, I didn't expect that.

Yeah, I was a little surprised.

> Doesn't this deactivate the feature detection? Eg. `IReadOnlySet` is only available for .NET 5+ and not for .NET Standard 2.0, this was detected before, but isn't in the new code. Not sure if this is an actual problem for Mapperly...

It could be an issue if Mapperly added a non backward compatible mapper for types implementing `IReadOnlySet`. I'm not too concerned by this, ImplementedTypes extracts a bunch of unused information